PDFBOX-2133: Tolerate reals in place of integers in Type 1 parser
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1602311 13f79535-47bb-0310-9956-ffa450edef68

+++ fontbox/src/main/java/org/apache/fontbox/type1/Token.java
@@ -99,7 +99,8 @@ class Token
 
     public int intValue()
     {
-        return Integer.parseInt(text);
+        // some fonts have reals where integers should be, so we tolerate it
+        return (int)Float.parseFloat(text);
     }
 
     public float floatValue()
